Add abbreviations
This is where things start getting weird. Some of these abbreviations are actually numbers (IV) some of these abbreviations are actually medical terms (IV), some are places (D.C.), some are scientific (DC), and some are names (NBC). This is good for now but we'll probably end up moving them later as we get more categories.